Add parsing of ZPT serial data and unify block comment styles
The ZPT unit outputs a serial data packet format with information
about received signal from a transmitter. If this is connected to Serial2
on the feather HUZZAH32 we can parse it and get informaiton such as
transmitter low battery or if the button has been learned by the
receiver and print up a status character int the top right to warn the user
'!' means low battery and '?' means unlearned button code.