turning.ino updates for logic, etc
Fix logic in not quite a state machine so states are not skipped
through (wasting a tick and causing a flicker on the screen) if they
are not going to be used (because the option is zero).

Fix logic if autonext is set on the last stage in a program.

ifdef out more ZPT serial sections if zpt serial is not used.

Make sure timer stops when running stops.

Comment and whitespace fixes